Integrando conexion con backend 1.0

+++ app/src/main/java/com/example/levelup/model/data/CarritoEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "carrito")
 data class CarritoEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val productoId: Int,
+    val productoId: Long,
     val nombre: String,
     val precio: Double,
     val cantidad: Int = 1,
