+++ app/src/main/java/com/example/levelup/remote/ProductosDTO.kt
@@ -3,7 +3,7 @@ package com.example.levelup.remote
 data class ProductosDTO(
     val id: Long,
     val nombre: String,
-    val precio: Double,
+    val precio: Long,
     val descripcion: String,
     val imagenUrl: String
 )
